cpuset: Implement hotreplug invariancy of cpusets configuration
The current cpusets are hotreplug-invariant only on cgroups v2
but, on legacy, they will act in a strange way: when a CPU is
hotunplugged, it will be removed from cpus_allowed mask, which
results in a loss of configuration data because, at hotreplug
time, that CPU will not be reinserted in the allowed list,
being it removed from both allowed and effective masks.

Implement invariancy of the cpusets configurations for
hotpluggging: now we correctly use the masks, allowing for
the EFFECTIVE CPUs mask to be not the same as the ALLOWED
CPUs mask.
If we allowed, for example, cpu 0 to 3 in our configuration
from userspace, now when we hotunplug the CPUs they will
still be allowed, but not usable.
When we hotreplug the CPU, it will be allowed and usable.

While at it, do the same for mems (even if we don't use it).

Signed-off-by: Samuel Pascua <pascua.samuel.14@gmail.com>