HID: logitech-hidpp: adapt to input_configured API change
The backport of an API change in the input subsystem clashed with
the backport of the logitec hid++ driver, as shown by this warning:

drivers/hid/hid-logitech-hidpp.c:838:22: warning: initialization from incompatible pointer type [-Wincompatible-pointer-types]

The result is an undefined return code when calling the input_configured
callback handler in this driver.

This backports the API change as well. This was originally part of
commit b2c68a2f1bab ("HID: hid-input: allow input_configured callback
return errors") but got skipped as the driver did not exist in
android-3.18 at that time.

Fixes: 7682bfa411f4 ("UPSTREAM: HID: hid-input: allow input_configured callback return errors")
Fixes: bd28e66ce1fb ("HID: Introduce hidpp, a module to handle Logitech hid++ devices")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>