BACKPORT: ANDROID: sched/fair: Select correct capacity state for energy_diff
The util returned from group_max_util is not capped at the max util
present in the group, so it can be larger than the capacity stored in
the array. Ensure that when this happens, we always use the last entry
in the array to fetch energy from.

Tested with synthetics on Juno board.

Fixes: b/38159576

(backported from android-4.4, Change-Id: I89fb52fb7e6)
Bug: 38159576
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: Icb3d7b02c6bbf302ee4eba95e6ff1216bbaf9b31