cpuquiet: Make the hotplug/isolation switch more robust
When switching hotplug<=>isolation, it's good to clean up
the whole state machine in the cpuquiet API, so that we
restart fresh.
This will make us to lose some cycles, but we embrace safer
switching operation for a very small cost.

Also, apply the same cpumask logic that has been applied to
rqbalance regarding the concept of "available CPUs" and
protect the switch operation with the cpu_lock mutex so
that we're extra sure that - even if this function gets
called without the main mutex protection - we are at least
protected against the main API worker.

Signed-off-by: Samuel Pascua <pascua.samuel.14@gmail.com>